moved the utility files to the ap/util directory
