+++ include/ap/argument_parser.hpp
@@ -11,7 +11,6 @@
 
 #include "argument.hpp"
 #include "detail/argument_token.hpp"
-#include "detail/concepts.hpp"
 #include "types.hpp"
 
 #include <algorithm>
