+++ include/ap/detail/argument_descriptor.hpp
@@ -9,9 +9,9 @@
 
 #pragma once
 
-#include "ap/detail/argument_name.hpp"
-#include "concepts.hpp"
-#include "str_utility.hpp"
+#include "ap/util/concepts.hpp"
+#include "ap/util/string.hpp"
+#include "argument_name.hpp"
 
 #include <cstdint>
 #include <format>
