+++ include/ap/util/concepts.hpp
@@ -3,7 +3,7 @@
 // Licensed under the MIT License. See the LICENSE file in the project root for full license information.
 
 /**
- * @file ap/detail/concepts.hpp
+ * @file ap/util/concepts.hpp
  * @brief Provides the general concept definitions.
  */
