+++ include/ap/util/typing.hpp
@@ -2,7 +2,10 @@
 // This file is part of the CPP-AP project (https://github.com/SpectraL519/cpp-ap).
 // Licensed under the MIT License. See the LICENSE file in the project root for full license information.
 
-/// @file ap/detail/typing_utility.hpp
+/**
+ * @file ap/util/typing.hpp
+ * @brief Provides common typing utility functions.
+ */
 
 #pragma once
