+++ tests/CMakeLists.txt
@@ -76,4 +76,4 @@ add_doctest("source/test_nargs_range.cpp")
 add_doctest("source/test_none_type_argument.cpp")
 add_doctest("source/test_optional_argument.cpp")
 add_doctest("source/test_positional_argument.cpp")
-add_doctest("source/test_str_utility.cpp")
+add_doctest("source/test_string_utility.cpp")
